# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Conecta Orto
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface RegistrationInput {
  name: string;
  email: string;
  phone: string;
  city: string;
  profession: string;
}

export interface EmailLookup {
  email: string;
}

export interface Registration {
  id: number;
  name: string;
  email: string;
  phone: string;
  city: string;
  profession: string;
  createdAt: string;
}

export interface RegistrationCount {
  count: number;
}

export interface EnrolledMinicourse {
  minicourseId: number;
  title: string;
  instructor: string;
}

export interface RegistrationWithEnrollments {
  id: number;
  name: string;
  email: string;
  phone: string;
  city: string;
  profession: string;
  createdAt: string;
  enrollments: EnrolledMinicourse[];
}

export interface MinicourseWithCount {
  id: number;
  title: string;
  instructor: string;
  description: string;
  duration: string;
  maxCapacity: number;
  enrollmentCount: number;
}

export interface EnrollmentInput {
  registrationId: number;
  minicourseId: number;
}

export interface Enrollment {
  id: number;
  registrationId: number;
  minicourseId: number;
  createdAt: string;
}

export interface AdminLoginInput {
  password: string;
}

export interface AdminToken {
  token: string;
}

export interface RegistrantWithEnrollments {
  id: number;
  name: string;
  email: string;
  phone: string;
  city: string;
  profession: string;
  createdAt: string;
  minicourses: string[];
}

export interface MinicourseEnrollmentStat {
  id: number;
  title: string;
  enrollmentCount: number;
  maxCapacity: number;
}

export interface AdminStats {
  totalRegistrants: number;
  minicourseEnrollments: MinicourseEnrollmentStat[];
}